git-svn-id: svn://192.168.1.10/public/CoffeeMud@25343 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/fakedb/backend/structure/FakeTable.java
@@ -775,7 +775,7 @@ public class FakeTable
 					addHole(prevRecord.offset, prevRecord.size);
 				final Integer holeOffset = takeHole(ofs);
 				if (holeOffset != null)
-					recordPos = holeOffset;
+					recordPos = holeOffset.intValue();
 				else
 				{
 					recordPos = fileSize;
